fix: Move service to core
