+++ docs/templates/package.json
@@ -37,7 +37,7 @@
     "branches": 80
   },
   "devDependencies": {
-    "@aerogear/app": "0.1.0",
+    "@aerogear/core": "0.1.0",
     "@types/chai": "4.1.2",
     "@types/chai-as-promised": "7.1.0",
     "@types/mocha": "^5.0.0",
@@ -56,6 +56,6 @@
     "typescript": "^2.8.1"
   },
   "peerDependencies": {
-    "@aerogear/app": "0.1.0"
+    "@aerogear/core": "0.1.0"
   }
 }
