+++ packages/app/package.json
@@ -26,15 +26,11 @@
   "homepage": "https://github.com/aerogear/aerogear-js-sdk#readme",
   "devDependencies": {
     "@types/chai": "^4.1.3",
-    "@types/lodash.find": "^4.6.3",
-    "@types/loglevel": "^1.5.3",
     "chai": "^4.1.2",
     "mocha": "^5.1.1",
     "nyc": "^11.7.1"
   },
   "dependencies": {
-    "@aerogear/core": "2.1.0",
-    "lodash.find": "^4.6.0",
-    "loglevel": "^1.6.1"
+    "@aerogear/core": "2.1.0"
   }
 }
