+++ packages/auth/package.json
@@ -49,7 +49,7 @@
     "branches": 80
   },
   "devDependencies": {
-    "@aerogear/app": "2.1.0",
+    "@aerogear/core": "2.1.0",
     "@types/chai": "4.1.2",
     "@types/chai-as-promised": "7.1.0",
     "@types/loglevel": "^1.5.3",
@@ -68,12 +68,11 @@
     "typescript": "^2.8.1"
   },
   "dependencies": {
-    "@aerogear/core": "2.1.0",
     "keycloak-js": "^4.8.3",
     "loglevel": "^1.6.1",
     "url": "^0.11.0"
   },
   "peerDependencies": {
-    "@aerogear/app": ">2.0.1"
+    "@aerogear/core": ">2.0.1"
   }
 }
