+++ packages/core/src/config/index.ts
@@ -1,2 +1,6 @@
+// API
 export { AeroGearConfiguration } from "./AeroGearConfiguration";
 export { ServiceConfiguration } from "./ServiceConfiguration";
+
+// Implementation
+export { ConfigurationService } from "./ConfigurationService";
