+++ packages/sync/package.json
@@ -31,7 +31,7 @@
     "access": "public"
   },
   "devDependencies": {
-    "@aerogear/app": "2.1.0",
+    "@aerogear/core": "2.1.0",
     "@types/chai": "^4.1.3",
     "@types/debug": "0.0.31",
     "@types/graphql": "^14.0.3",
@@ -44,7 +44,7 @@
     "typescript": "^3.1.6"
   },
   "peerDependencies": {
-    "@aerogear/app": ">2.0.1"
+    "@aerogear/core": ">2.0.1"
   },
   "dependencies": {
     "@aerogear/core": "2.1.0",
